pyproject.toml
[project]
name = "project-name"
description = ""
version = "1.0.0"
authors = [{ name = "davyddd" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "uvicorn[standard]==0.32.0",
    "fastapi==0.118.1",
    "typer==0.21.1",
    "pydantic[email]==2.12.5",
    "pydantic-settings==2.6.0",
    "dddesign==1.1.10",
    "sentry-sdk[fastapi]==2.17.0",
    "psycopg[binary]==3.2.3",
    "sqlmodel==0.0.22",
    "alembic==1.13.3",
    "dramatiq[redis,watch,prometheus]==2.0.1",
    "apscheduler==3.10.4",
    "clickhouse-connect==0.8.16",
    "httpx==0.28.1",
    "python-json-logger==3.3.0",
    "ddutils==0.1.4",
    "clickhouse-sqlalchemy==0.3.2",
    "jinja2==3.1.6",
    "ddsql==0.0.2",
    "redis==6.4.0",
    "aiokafka==0.13.0",
    "python-multipart==0.0.22",
]

[dependency-groups]
dev = [
    "ipdb==0.13.9",
    "ipython==8.3.0",
    "ty==0.0.15",
    "ruff==0.1.7",
    "import-linter==1.12.1",
    "complexipy==5.2.0",
]

[tool.uv]
package = false
